move api handlers to non-pro
