+++ services/src/api/handlers/machine_learning.rs
@@ -157,7 +157,7 @@ mod tests {
 
         let body = vec![(
             "model.onnx",
-            include_bytes!("../../../../../test_data/pro/ml/onnx/test_classification.onnx"),
+            include_bytes!("../../../../test_data/pro/ml/onnx/test_classification.onnx"),
         )];
 
         let req = test::TestRequest::post()
