+++ services/src/api/handlers/mod.rs
@@ -8,12 +8,15 @@ use std::str::FromStr;
 pub mod datasets;
 pub mod ebv;
 pub mod layers;
+pub mod machine_learning;
+pub mod permissions;
 pub mod plots;
 pub mod projects;
 pub mod session;
 pub mod spatial_references;
 pub mod tasks;
 pub mod upload;
+pub mod users;
 pub mod wcs;
 pub mod wfs;
 pub mod wms;
