--- services/src/pro/api/cli/mod.rs
@@ -1,3 +1,0 @@
-mod add_from_directory;
-
-pub use add_from_directory::add_datasets_from_directory;
