--- services/src/pro/api/handlers.rs
@@ -1,4 +1,0 @@
-pub mod datasets;
-pub mod machine_learning;
-pub mod permissions;
-pub mod users;
