--- services/src/pro/api/mod.rs
@@ -1,5 +1,0 @@
-mod apidoc;
-pub mod cli;
-pub mod handlers;
-
-pub use apidoc::ApiDoc;
