+++ services/src/pro/mod.rs
@@ -1,6 +1,5 @@
 // This is an inclusion point of Geo Engine Pro – TODO: migrate
 
-pub mod api;
 pub mod contexts;
 pub mod datasets;
 pub mod layers;
