+++ services/src/pro/users/mod.rs
@@ -5,7 +5,7 @@ mod user;
 mod userdb;
 
 pub(crate) use oidc::OidcError;
-pub(super) use oidc::{AuthCodeRequestURL, AuthCodeResponse, OidcDisabled, OidcManager};
+pub(crate) use oidc::{AuthCodeRequestURL, AuthCodeResponse, OidcDisabled, OidcManager};
 #[cfg(test)]
 pub(super) use oidc::{DefaultJsonWebKeySet, DefaultProviderMetadata, OidcTokens, UserClaims};
 pub use session::{UserInfo, UserSession};
